implement database CRUD API

+++ src/main/java/com/example/nosql_database_management_system/DAO/DatabaseDAO.java
@@ -1,5 +1,7 @@
 package com.example.nosql_database_management_system.DAO;
 
+import com.example.nosql_database_management_system.exception.ForbiddenException;
+import com.example.nosql_database_management_system.exception.ResourceNotFoundException;
 import org.springframework.stereotype.Repository;
 
 import java.io.File;
@@ -10,20 +12,30 @@ import java.util.List;
 public class DatabaseDAO {
     private final String BASE_PATH = "databases/";
 
-
+    public DatabaseDAO() {
+        File base = new File(BASE_PATH);
+        if (!base.exists()) {
+            base.mkdir();
+        }
+    }
     public void createDB(String name) {
         File DBpath = new File(BASE_PATH + name);
-
         if (!DBpath.exists()) {
             DBpath.mkdir();
 
             File schemaPath = new File(BASE_PATH + name + "/schemas");
             schemaPath.mkdir();
         }
+        else {
+            throw new ForbiddenException("database is already existed");
+        }
     }
 
     public void deleteDB(String name) {
         File file = new File(BASE_PATH + name);
+        if (!file.exists()) {
+            throw new ResourceNotFoundException("Database not found: " + name);
+        }
         deleteDirectory(file);
     }
 
@@ -39,7 +51,7 @@ public class DatabaseDAO {
         file.delete();
     }
 
-    public List<String> listDBs() {
+    public List<String> getAllDBs() {
         File folder = new File(BASE_PATH);
         String[] names = folder.list();
